+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 node_modules/
 .env
 package-lock.json
+coverage
+.DS_Store
