+++ index.js
@@ -1,29 +1,11 @@
 require("dotenv").config();
+const config = require("./config.js");
+const cache = require("./cache.js");
+const buildServer = require("./server/build.js");
 
 const mongoose = require("mongoose");
 
-const fastify = require("fastify");
-const cors = require("@fastify/cors");
-const config = require("./config.js");
-const cache = require("./cache.js");
-
-
-const mainRoutes = require("./routes/urls");
-const cacheRoutes = require("./routes/cache");
-
-
-const app = fastify({
-  trustProxy: true,
-});
-app.register(cors);
-
-mainRoutes(app);
-cacheRoutes(app);
-
-
-app.decorate("mongoose", mongoose);
-app.decorate("cache", cache);
-
+const app = buildServer();
 
 // DB connection
 mongoose
