+++ mongo-connection.js
@@ -6,7 +6,6 @@ const connectDB = async () => {
   try {
     await mongoose.connect(mongoUrl);
 
-
     console.log("MongoDB connected!!");
   } catch (err) {
     console.log("Failed to connect to MongoDB", err);
