+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node index.js",
     "dev": "npx nodemon --exec \"node index.js\"",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "vitest run --coverage",
     "lint": "npx rome check --apply-suggested .",
     "format": "npx rome format --write ."
   },
@@ -29,6 +29,8 @@
   },
   "devDependencies": {
     "nodemon": "^2.0.20",
-    "rome": "^11.0.0"
+    "rome": "^11.0.0",
+    "vitest": "^0.28.4",
+    "@vitest/coverage-c8": "^0.28.4"
   }
 }
