+++ routes/urls.js
@@ -46,14 +46,12 @@ module.exports = function (app) {
         return cache.getCache().get(cacheKey);
       }
 
-
       await checkConnection();
       if (endIndex < (await Yardim.countDocuments().exec())) {
         results.next = {
           page: page + 1,
           limit,
         };
-
       }
 
       if (startIndex > 0) {
@@ -67,7 +65,6 @@ module.exports = function (app) {
 
       results.totalPage = Math.ceil((await Yardim.countDocuments(query)) / limit);
       results.data = await Yardim.find(query).sort({ _id: -1 }).limit(limit).skip(startIndex).exec();
-
 
       results.data = results.data.map((yardim) => {
         yardim.telefon = yardim.telefon.replace(/.(?=.{4})/g, "*");
@@ -84,7 +81,6 @@ module.exports = function (app) {
           });
         }
         return yardim;
-
       });
 
       cache.getCache().set(cacheKey, results);
@@ -117,7 +113,6 @@ module.exports = function (app) {
               minLength: 10,
               maxLength: 11,
               pattern: "^d+$",
-              
             },
           },
           required: ["yardimTipi", "adSoyad", "adres", "acilDurum"],
@@ -157,13 +152,11 @@ module.exports = function (app) {
 
       const fields = {};
 
-
       // TODO: Bunlarin hepsini JSON schema'ya tasiyalim.
       for (const key in req.body) {
         if (key.startsWith("fields-")) {
           const fieldName = key.split("-")[1];
           fields[fieldName] = req.body[key];
-
         }
       }
 
@@ -314,7 +307,6 @@ module.exports = function (app) {
           page: page + 1,
           limit,
         };
-
       }
 
       if (startIndex > 0) {
@@ -389,12 +381,10 @@ module.exports = function (app) {
         };
       }
 
-
       if (location) {
         query = {
           $and: [query, { sehir: location }],
         };
-
       }
 
       if (dest) {
@@ -465,13 +455,11 @@ module.exports = function (app) {
         ],
       };
 
-
       if (helpType) {
         query = {
           $and: [query, { yardimTipi: helpType }],
         };
       }
-
 
       if (vehicle) {
         query = {
@@ -479,12 +467,10 @@ module.exports = function (app) {
         };
       }
 
-
       if (yardimDurumuQuery) {
         query = {
           $and: [query, { yardimDurumu: yardimDurumuQuery }],
         };
-
       }
 
       if (acilDurumQuery) {
@@ -541,7 +527,6 @@ module.exports = function (app) {
       yardimKaydi: yardimKaydi,
     });
     if (!results) {
-
       res.statusCode = 404;
       return { status: 404 };
     }
@@ -551,7 +536,6 @@ module.exports = function (app) {
       yardimKaydi,
     };
   });
-
 
   app.get("/yardimet/:id", async (req, res) => {
     const cacheKey = `yardimet_${req.params.id}`;
@@ -570,7 +554,6 @@ module.exports = function (app) {
     }
     cache.getCache().set(cacheKey, results);
     if (!results) {
-
       res.statusCode = 404;
       return { status: 404 };
     }
@@ -600,7 +583,6 @@ module.exports = function (app) {
         adSoyad: req.body.adSoyad,
         email: req.body.email,
         mesaj: req.body.mesaj,
-
       });
 
       if (existingIletisim) {
@@ -685,7 +667,6 @@ module.exports = function (app) {
     },
   );
 };
-
 
 async function checkConnection() {
   try {
